+++ logging/kit/ctxkit/doc.go
@@ -1,0 +1,14 @@
+/*
+`ctxkit` is a ctxlogger that is backed by go-kit
+
+It accepts a user-configured `log.Logger` that will be used for logging. The same `log.Logger` will
+be populated into the `context.Context` passed into gRPC handler code.
+
+You can use `ctxkit.Extract` to log into a request-scoped `log.Logger` instance in your handler code.
+
+As `ctxkit.Extract` will iterate all tags on from `grpc_ctxtags` it is therefore expensive so it is advised that you
+extract once at the start of the function from the context and reuse it for the remainder of the function (see examples).
+
+Please see examples and tests for examples of use.
+*/
+package ctxkit